+++ lib/screens/mapPage/ExploreScreen.js
@@ -584,10 +584,10 @@ const [state, setState] = useState({
       </View>
     </View>
   </View>
-))}
+))} 
 
       </Animated.ScrollView>
-      <NoteDetailModal 
+    <NoteDetailModal 
       isVisible={isModalVisible} 
       onClose={() => setModalVisible(false)} 
       note={selectedNote}
